rename tokensetitem, remove size
